+++ doc/ReleaseNotes.md
@@ -1,6 +1,22 @@
-﻿### 1.1.8
+﻿﻿### 1.1.12
 
  * Ignore System.EnterpriseServices.Thunk.dll, see https://github.com/matthid/Yaaf.FSharp.Scripting/issues/4
+
+﻿### 1.1.11
+
+ * Improve logging output.
+ 
+ ### 1.1.10
+
+ * Improve logging output.
+
+### 1.1.9
+
+ * Bugfix: Correctly log large argument lists.
+
+### 1.1.8
+
+ * Add logging with System.Diagnostics.TraceSource (configure in app.config with the name "Yaaf.FSharp.Scripting")
 
 ### 1.1.7